+++ server/src/views/translations/ai-translate-all/procedures/llm-translate.ts
@@ -1,7 +1,7 @@
 import OpenAI from "openai";
 import { zodResponseFormat } from "openai/helpers/zod";
 import { z } from "zod";
-import { env } from "../../../config";
+import { env } from "../../../../config";
 
 const openaiClient = new OpenAI({ apiKey: env.OPENAI_API_KEY });
